refactor(AboutPage): split into modular components for better maintainability
